feat : app navigator component added
