+++ App.tsx
@@ -1,11 +1,6 @@
 import { useState } from "react";
 import { useDeviceOrientation } from "@react-native-community/hooks";
 import { NavigationContainer } from "@react-navigation/native";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { MaterialCommunityIcons } from "@expo/vector-icons";
-
-import Screen from "./app/components/Screen";
-import AuthNavigation from "./app/navigation/AuthNavigation";
 
 import { AppTextInput } from "./app/components/TextInput";
 import Card from "./app/components/Card";
@@ -15,41 +10,18 @@ import WelcomeScreen from "./app/screens/WelcomeScreen";
 import LoginScreen from "./app/screens/LoginScreen";
 import ViewImageScreen from "./app/screens/ViewImageScreen";
 import MessagesScreen from "./app/screens/MessageScreen";
-import DetailsScreen from "./app/screens/ListingDetailsScreen";
+import ListingDetailsScreen from "./app/screens/ListingDetailsScreen";
 import ListingEditScreen from "./app/screens/ListingEditScreen";
 import ListingsScreen from "./app/screens/ListingsScreen";
 import AccountScreen from "./app/screens/AccountScreen";
 
 import NavigationTheme from "./app/navigation/NavigationTheme";
+import AuthNavigation from "./app/navigation/AuthNavigation";
+import AppNavigator from "./app/navigation/AppNavigator";
+import Screen from "./app/components/Screen";
 import colors from "./app/config/colors";
 
 interface IDoneWithItProps {}
-
-const Tab = createBottomTabNavigator();
-
-const TabNavigator: React.FunctionComponent = () => {
-  return (
-    <Tab.Navigator
-      screenOptions={{
-        tabBarActiveBackgroundColor: "tomato",
-        tabBarActiveTintColor: "white",
-        tabBarInactiveBackgroundColor: "#eee",
-        tabBarInactiveTintColor: "black",
-      }}
-    >
-      <Tab.Screen
-        options={{
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons name='home' color={color} size={size} />
-          ),
-        }}
-        name='Home'
-        component={WelcomeScreen}
-      />
-      <Tab.Screen name='Login' component={LoginScreen} />
-    </Tab.Navigator>
-  );
-};
 
 const DoneWithIt: React.FunctionComponent<IDoneWithItProps> = (props) => {
   // const categories = [
@@ -65,10 +37,7 @@ const DoneWithIt: React.FunctionComponent<IDoneWithItProps> = (props) => {
 
   return (
     <NavigationContainer theme={NavigationTheme}>
-      {/* <ListingsScreen/> */}
-      {/* <AuthNavigation /> */}
-      {/* <TabNavigator /> */}
-      {/* <AccountScreen /> */}
+      <AppNavigator />
     </NavigationContainer>
   );
 };
