+++ app/navigation/AppNavigator.tsx
@@ -1,0 +1,53 @@
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import { MaterialCommunityIcons } from "@expo/vector-icons";
+
+import ListingEditScreen from "../screens/ListingEditScreen";
+import ListingsScreen from "../screens/ListingsScreen";
+import AccountScreen from "../screens/AccountScreen";
+
+interface IAppNavigatorProps {}
+
+const Tab = createBottomTabNavigator();
+
+const AppNavigator: React.FunctionComponent<IAppNavigatorProps> = () => {
+  return (
+    <Tab.Navigator
+      screenOptions={{
+        tabBarActiveBackgroundColor: "tomato",
+        tabBarActiveTintColor: "white",
+        tabBarInactiveBackgroundColor: "#eee",
+        tabBarInactiveTintColor: "black",
+      }}
+    >
+      <Tab.Screen
+        options={{
+          tabBarIcon: ({ color, size }) => (
+            <MaterialCommunityIcons name='home' color={color} size={size} />
+          ),
+        }}
+        name='Listing'
+        component={ListingsScreen}
+      />
+      <Tab.Screen
+        options={{
+          tabBarIcon: ({ color, size }) => (
+            <MaterialCommunityIcons name='home' color={color} size={size} />
+          ),
+        }}
+        name='Listing Edit'
+        component={ListingEditScreen}
+      />
+      <Tab.Screen
+        options={{
+          tabBarIcon: ({ color, size }) => (
+            <MaterialCommunityIcons name='home' color={color} size={size} />
+          ),
+        }}
+        name='Account'
+        component={AccountScreen}
+      />
+    </Tab.Navigator>
+  );
+};
+
+export default AppNavigator;
